requirements.txt
# FastAPI & Server
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.28
psycopg2-binary==2.9.9    # PostgreSQL driver (works with Neon)
alembic==1.13.1           # DB migrations

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data Validation
pydantic==2.6.3
pydantic-settings==2.2.1

# CORS & Middleware
starlette==0.36.3

# Caching / Optional
redis==5.0.3

# AI (Gemini)
google-generativeai==0.3.2

# Utilities
httpx==0.27.0             # async HTTP client
python-dotenv==1.0.1      # for env variables (.env file)

#Email Validator 
email-validator